+++ extensions/iceworks-project-creator/src/extension.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { connectService, getHtmlForWebview } from '@iceworks/vscode-webview/lib/vscode';
-import { initExtensionConfiguration, Logger } from '@iceworks/common-service';
+import { initExtension, Logger } from '@iceworks/common-service';
 import services from './services/index';
 
 // eslint-disable-next-line
@@ -14,24 +14,18 @@ export function activate(context: vscode.ExtensionContext) {
   // data collection
   const logger = new Logger(name, globalState);
   logger.recordDAU();
-  logger.recordOnce({
-    module: 'main',
-    action: 'activate',
-    data: {
-      version,
-    }
-  });
+  logger.recordActivate(version);
 
   // auto set configuration
-  initExtensionConfiguration(globalState);
+  initExtension(context);
 
   function activeWebview() {
-    const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', '创建应用', ViewColumn.One, {
+    const webviewPanel: vscode.WebviewPanel = window.createWebviewPanel('iceworks', '创建应用 - Iceworks', ViewColumn.One, {
       enableScripts: true,
       retainContextWhenHidden: true,
     });
     webviewPanel.webview.html = getHtmlForWebview(extensionPath);
-    connectService(webviewPanel.webview, subscriptions, { services, logger });
+    connectService(webviewPanel, context, { services, logger });
   }
 
   subscriptions.push(vscode.commands.registerCommand('iceworks-project-creator.start', function () {