+++ extensions/iceworks-project-creator/web/src/pages/CreateProject/components/ScaffoldMarket/index.tsx
@@ -1,25 +1,19 @@
 /* eslint-disable */
 import React, { useState, useEffect } from 'react';
-import { Collapse, Notification, Loading, Button, Icon } from '@alifd/next';
-import SelectCard from '@/components/SelectCard';
+import { Collapse, Notification, Loading, Button, Icon, Divider } from '@alifd/next';
+import MobileScaffoldCard from '@/components/MobileScaffoldCard';
+import ScaffoldCard from '@/components/ScaffoldCard';
 import NotFound from '@/components/NotFound';
 import callService from '@/callService';
 import { IMaterialSource, IMaterialScaffold } from '@iceworks/material-utils';
+import { mainScaffoldsList, tsScaffoldsList, jsScaffoldsList } from '@/constant';
 import styles from './index.module.scss';
 
-const tsScaffolds = [
-  '@alifd/fusion-design-pro',
-  '@alifd/scaffold-lite',
-]
-const jsScaffolds = [
-  '@alifd/fusion-design-pro-js',
-  '@alifd/scaffold-lite-js'
-]
+const projectTypes = ['react', 'rax', 'vue'];
 
-const ScaffoldMarket = ({ onScaffoldSelect, children, onOpenConfigPanel }) => {
-  const [materialSourceSelected, setMaterialSourceSelected] = useState<any>({});
-  const [materialSelected, setMaterialSelected] = useState(null);
-  const [materialSources, setMaterialSources] = useState<Array<IMaterialSource>>([]);
+const ScaffoldMarket = ({ onScaffoldSelect, children, onOpenConfigPanel, materialSources }) => {
+  const [selectedSource, setSelectedSource] = useState<any>({});
+  const [SelectedMaterial, setSelectedMaterial] = useState(null);
   const [mainScaffolds, setMainScaffolds] = useState<IMaterialScaffold[]>([]);
   const [otherScaffolds, setOtherScaffolds] = useState<IMaterialScaffold[]>([]);
   const [loading, setLoading] = useState<boolean>(false);
@@ -27,12 +21,12 @@ const ScaffoldMarket = ({ onScaffoldSelect, children, onOpenConfigPanel }) => {
   async function onMaterialSourceClick(scaffold: IMaterialSource) {
     try {
       setLoading(true);
-      setMaterialSourceSelected(scaffold);
+      setSelectedSource(scaffold);
       const data = await getScaffolds(scaffold.source);
       const { mainScaffolds, otherScaffolds } = data as any;
       setMainScaffolds(mainScaffolds);
       setOtherScaffolds(otherScaffolds);
-      setMaterialSelected(null);
+      setSelectedMaterial(null);
     } catch (err) {
       // ignore
     } finally {
@@ -41,22 +35,19 @@ const ScaffoldMarket = ({ onScaffoldSelect, children, onOpenConfigPanel }) => {
   }
 
   function onScaffoldMaterialClick(scaffold) {
-    setMaterialSelected(scaffold.name);
+    setSelectedMaterial(scaffold.name);
     onScaffoldSelect(scaffold);
-  }
-
-  async function getScaffoldSources() {
-    const materialSources: any = await callService('material', 'getSources') as IMaterialSource[];
-    setMaterialSources(materialSources);
-    setMaterialSourceSelected(materialSources[0]);
-    return materialSources;
   }
 
   async function getScaffolds(source: string) {
     try {
       const scaffolds = await callService('project', 'getScaffolds', source) as IMaterialScaffold[];
-      const main = scaffolds.filter(scaffold => !jsScaffolds.includes(scaffold.source.npm));
-      const other = scaffolds.filter(scaffold => jsScaffolds.includes(scaffold.source.npm));
+      let main = scaffolds.filter(scaffold => mainScaffoldsList.includes(scaffold.source.npm));
+      let other = scaffolds.filter(scaffold => !mainScaffoldsList.includes(scaffold.source.npm));
+      if (!main.length && other.length) {
+        main = other;
+        other = [];
+      }
       return { mainScaffolds: main, otherScaffolds: other }
     } catch (e) {
       Notification.error({ content: e.message });
@@ -67,7 +58,10 @@ const ScaffoldMarket = ({ onScaffoldSelect, children, onOpenConfigPanel }) => {
   async function initData() {
     setLoading(true);
     try {
-      const materialSources = await getScaffoldSources();
+      if (!materialSources.length) {
+        return;
+      }
+      setSelectedSource(materialSources[0]);
       const source = materialSources[0].source;
 
       const data = await getScaffolds(source);
@@ -81,57 +75,57 @@ const ScaffoldMarket = ({ onScaffoldSelect, children, onOpenConfigPanel }) => {
     }
   }
 
-  async function onRefreshMaterialSource() {
-    await initData();
-  }
-
   useEffect(() => {
     initData();
-  }, []);
+  }, [materialSources]);
   return (
     <div className={styles.container}>
       <div className={styles.content}>
         <div className={styles.scaffoldsSource}>
-          <div className={styles.refreshBtn}>
-            <Button text onClick={onRefreshMaterialSource}><Icon type="refresh" />刷新</Button>
-          </div>
           <div className={styles.sourcesList}>
-            {materialSources && materialSources.map(item => (
-              <SelectCard
+            {materialSources && materialSources.map(item => {
+              const projectType = item.type.toLocaleLowerCase();
+              return <ScaffoldCard
                 key={item.name}
-                title={item.name}
-                content={<span className={styles.userSelect}>{item.description}</span>}
-                selected={materialSourceSelected['name'] && materialSourceSelected.name === item.name}
-                style={{ width: 160 }}
+                title={
+                  <div className={styles.cardTitle}>
+                    {<img src={require(`@/assets/${projectTypes.includes(projectType) ? projectType + '.svg' : 'logo.png'}`)} alt="projectType" width={25} height={25} />}
+                    <div >{item.name}</div>
+                  </div>
+                }
+                content={item.description}
+                selected={selectedSource.name && selectedSource.name === item.name}
+                style={{ width: 160, height: 110 }}
                 onClick={() => onMaterialSourceClick(item)}
               />
-            ))}
+            })}
           </div>
           <div className={styles.addSource}>
             <Button className={styles.btn} onClick={onOpenConfigPanel}><Icon type="add" /></Button>
           </div>
         </div>
-
+        <Divider direction="ver" style={{ height: '100%' }} />
         <div className={styles.scaffolds}>
           {loading ? <Loading visible={loading} className={styles.loading} /> : <>
             <div className={styles.mainScaffolds}>
               {!!mainScaffolds.length ? mainScaffolds.map(item => {
-                const scaffoldType = tsScaffolds.includes(item.source.npm) ? 'ts' :
-                  jsScaffolds.includes(item.source.npm) ? 'js' :
-                    ''
+                const scaffoldType = tsScaffoldsList.includes(item.source.npm) ? 'ts' :
+                  jsScaffoldsList.includes(item.source.npm) ? 'js' :
+                    '';
+                const isRax = selectedSource.type === 'rax';
+                const CardComponent = isRax ? MobileScaffoldCard : ScaffoldCard;
                 return (
-                  <SelectCard
+                  <CardComponent
                     key={item.name}
                     title={
-                      <div>
-                        {scaffoldType && <img className={styles.cardProjectType} src={require(`@/assets/${scaffoldType}.svg`)} alt="projectType" width={20} height={20} />}
-                        <span className={styles.cardTitle}>{item.title.replace(' - TS', '')}</span>
+                      <div className={styles.cardTitle}>
+                        {scaffoldType && <img src={require(`@/assets/${scaffoldType}.svg`)} alt="languageType" width={20} height={20} />}
+                        <div>{item.title.replace(' - TS', '')}</div>
                       </div>
                     }
-                    content={<span className={styles.userSelect}>{item.description}</span>}
-                    media={<img height={120} src={item.screenshot} alt="screenshot" style={{ padding: '10px 10px 0' }} />}
-                    selected={materialSelected === item.name}
-                    style={{ width: 190, height: 250 }}
+                    content={item.description}
+                    media={item.screenshot}
+                    selected={SelectedMaterial === item.name}
                     onClick={() => onScaffoldMaterialClick(item)}
                   />
                 )
@@ -143,32 +137,31 @@ const ScaffoldMarket = ({ onScaffoldSelect, children, onOpenConfigPanel }) => {
               <Collapse.Panel title="查看更多">
                 <div className={styles.collapseScaffolds}>
                   {otherScaffolds.map(item => {
-                    const scaffoldType = tsScaffolds.includes(item.source.npm) ? 'ts' :
-                      jsScaffolds.includes(item.source.npm) ? 'js' :
-                        ''
+                    const scaffoldType = tsScaffoldsList.includes(item.source.npm) ? 'ts' :
+                      jsScaffoldsList.includes(item.source.npm) ? 'js' :
+                        '';
+                    const isRax = selectedSource.type.toLocaleLowerCase() === 'rax';
+                    const CardComponent = isRax ? MobileScaffoldCard : ScaffoldCard;
                     return (
-                      <SelectCard
+                      <CardComponent
                         key={item.name}
                         title={
-                          <div>
-                            {scaffoldType && <img className={styles.cardProjectType} src={require(`@/assets/${scaffoldType}.svg`)} alt="projectType" width={20} height={20} />}
-                            <span className={styles.cardTitle}>{item.title.replace(' - JS', '')}</span>
+                          <div className={styles.cardTitle}>
+                            {scaffoldType && <img src={require(`@/assets/${scaffoldType}.svg`)} alt="languageType" width={20} height={20} />}
+                            <div>{item.title.replace(' - JS', '')}</div>
                           </div>
                         }
-                        content={<span className={styles.userSelect}>{item.description}</span>}
-                        media={<img height={120} src={item.screenshot} alt="screenshot" style={{ padding: '10px 10px 0' }} />}
-                        selected={materialSelected === item.name}
-                        style={{ width: 190, height: 250 }}
+                        content={item.description}
+                        media={item.screenshot}
+                        selected={SelectedMaterial === item.name}
                         onClick={() => onScaffoldMaterialClick(item)}
                       />
                     )
                   })}
                 </div>
               </Collapse.Panel>
-            </Collapse>
-            }
-          </>
-          }
+            </Collapse>}
+          </>}
         </div>
       </div>
       <div className={styles.action}>