+++ extensions/iceworks-style-helper/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 0.3.0
+
+Add Sass Variables Helper:
+
+- Show the value of sass variables when hover.
+- Automatic completion When editing the sass variables.
+
 ## 0.2.0
 
 Add `class` name completions.